Set mode and owner on purify_serializer_path
When `caUtils clean-caches` runs it deletes everything inside the temp
directory including the HTMLPurifier serialiser path.
That was fixed in 373ca85826506be0f6c6466caf9d9b9f71ada5e5 but the
directory is created with the permissions of the user running the
script.
This commit adds some extra steps to ensure the `purify_serializer_path`
permissions and ownership match the parent tmp directory.